reorganization for V1
many minor updates:
I deleted some commented sections.
reorganized some of the plots.

creature - fixed some issue of the eyes (it was hardcoded to work in the way it did and had some redundancies).
also moved eat and consume energy to the creature module.
the dynamic height and weight now starts low and grow as the creature eats.

environment - the extent is now computed once and stored instead of at every any iteration.

static_traits - moved the id to this module instead of in creature.

simulation - some minor changes +
+ moved some assignments to the classes
+ added the option for different boundary conditions (zero or mirror speed [the previous treatment of zero speed gave bias to move right])
+fixed eyes problem - it was a single eye instead of list eyes (actually a list of one eye).
refactored the list of reproduce and die to be more consistent.
fixed an error with the new plot (the placeholders had an error).
the focused agent is now marked better (a stand out circle instead of a star)

config - reordered so it will be clean and readable