Update graphic_brain.py
refactored internal to hidden for readablity.
mutate: created the mutate method and updated it to forget parent's experience.
plot method: updated the visuals so it won't recalculate the structure with each plot and added the ability to give it axis.
general: refactored it all to automatically detect the name base on the node type