Lineage and traits graphs (#11)
Physics model update, added lineage plot and trait plots in the animation.
__________________________________________________________________________________

* initial rebalance

it seems to work better now

* some minor plot fixes

to fix the set_X  commands and to make room for the force and energy debugging

* adjusted plots and some more rebalancing

* Minor changes

The energy of each grass is now much smaller but more are generated - it cause the creature to explore more.

* fixed lineage_tree.py to be able to start in the middle based on given step.

* changed kinetic force + added traits histogram and scatter

* changed drag force to depend on height also

---------

Co-authored-by: SnirNehemia <112329287+SnirNehemia@users.noreply.github.com>