Balance and acceleration - full grass KDtree support and adolescence (#12)
* initial rebalance

it seems to work better now

* some minor plot fixes

to fix the set_X  commands and to make room for the force and energy debugging

* adjusted plots and some more rebalancing

* Minor changes

The energy of each grass is now much smaller but more are generated - it cause the creature to explore more.

* added some features:

* (graphic_brain, physical_model) - Energy consumption based on brain size
* Refactored and modified mutation to normal distribution. Now mutation config is std instead of maximal variation.
* Environment - Fixed a minor bug in grass generation
* Added adolescence - now in the first max_age*adolescense_fraction of a creature life it can grow but not reproduce and in the rest of its life it's vice versa.

* Full grass KDtree support

the changes are:
* now we only seek grass once (using the KDtree) and save the results for each creature in a dictionary
* we then use the same list for the brain input and for the food search (for eating).
* then, if a creature eats an item, we add it to a list and in the rest of this step treats it as if it is gone.
* finally, we remove all the items that were eaten in the end of the step and update the KDtree.
This process ensure that the KDtree is up to date in the sense that throughout the step the grass indices remain unchanged.
This change is in simulation and environment