Physics plots (#19)
* initial rebalance

it seems to work better now

* some minor plot fixes

to fix the set_X  commands and to make room for the force and energy debugging

* adjusted plots and some more rebalancing

* Minor changes

The energy of each grass is now much smaller but more are generated - it cause the creature to explore more.

* added some features:

* (graphic_brain, physical_model) - Energy consumption based on brain size
* Refactored and modified mutation to normal distribution. Now mutation config is std instead of maximal variation.
* Environment - Fixed a minor bug in grass generation
* Added adolescence - now in the first max_age*adolescense_fraction of a creature life it can grow but not reproduce and in the rest of its life it's vice versa.

* Full grass KDtree support

the changes are:
* now we only seek grass once (using the KDtree) and save the results for each creature in a dictionary
* we then use the same list for the brain input and for the food search (for eating).
* then, if a creature eats an item, we add it to a list and in the rest of this step treats it as if it is gone.
* finally, we remove all the items that were eaten in the end of the step and update the KDtree.
This process ensure that the KDtree is up to date in the sense that throughout the step the grass indices remain unchanged.
This change is in simulation and environment

* minor time\step fixes

mainly in the plot desplay

* friction is more elegant now and more step-sec corrections

* fixes a mistake in the static friction test - it is negative

* added the creature_log module

The plots are still in creature. it will be changed in the next commit.
The simulation still runs.

* intorucing plot_utils + minor bug fix and order

We need to decide on order of log\age update - now it's not consistent for children and founding generation.

* closed loop in brain and plot correction

now the closed loop is detected and a sigmoid activation is automatically applied there to avoid exploding number

* added plot force support

* animation axes rework

reordered axes for animation plot to be more elegant

* save record for all creatures

up until now some of it was for agents only

* plot games and conflict stuff

* angle plot is now in rad/pi

* fully working parametric dashboard

I still need to test if with structures and organize the code, it's a bit of a mess

* minor layout fixes

* structure support for plor_balance_struct_...

struct_checkpoint now runs with two structs
I need to check it further as I build the dashboard

* Update requirements.txt

* Update struct_checkpoint_works - now sliders works

now it should work.
all that is left is to add more subplots and rearrange the subplots UI.

* first attempt at energy over time balance plot

* minor correction

* fully working balance dashboard

all that is left to do is organize the code for readability and do the actual balance

* now everything is working!

And it's very easy to use.

* Organized it in one main function

* rotated the force back and refactoring

I previously rotated the propulsion force and refactored something for convenience. now I changed it back.
also fixed a double record of speed in simulation and creature

* both plot_balance and simulation works

I fixed some double entries in the logs and some minor errors that prevented the simulation from running.
Now all that's left is to make to create and call the functions directly from the creature class.

* removed creature dependency on config Brain

and got some interesting run from this config

* cosmetics

* Delete checkpoint.py

Snir said it's not required.

* Update creature.py

removed lines in comment

* Update creature.py

removed comments.

* Update plot_utils.py

fixed imported config and physical model.

* Update simulation_utils.py

Snir is idiot

---------

Co-authored-by: SaarNehemia <51112173+SaarNehemia@users.noreply.github.com>